% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_homework.R
\name{launch_homework}
\alias{launch_homework}
\title{Launch Rmd files for homework}
\usage{
launch_homework(class = 1, show_answers = FALSE, testing = FALSE)
}
\arguments{
\item{class}{enter the relevant class number (as an int)}

\item{show_answers}{enter TRUE or FALSE (default is FALSE)}

\item{testing}{enter TRUE or FALSE (default is FALSE)}
}
\value{
opens the corresponding Rmd file in RStudio

opens the corresponding Rmd file WITH ANSWERS in RStudio

appropriate path is selected depending on what environment the package is run in. USERS CAN IGNORE
}
\description{
This function launches the named class Rmd file
}
\author{
Akshay Bareja
}
