% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_project.R
\name{launch_project}
\alias{launch_project}
\title{Launch Rmd files for in-class activities}
\usage{
launch_project(class = 1, show_answers = FALSE, testing = FALSE)
}
\arguments{
\item{class}{enter the relevant class number (as an int). Opens the corresponding Rmd file in RStudio.}

\item{show_answers}{enter TRUE or FALSE (default is FALSE). Opens the corresponding Rmd file WITH ANSWERS in RStudio.}

\item{testing}{enter TRUE or FALSE (default is FALSE). Appropriate path is selected depending on what environment the package is run in. USERS CAN IGNORE.}
}
\description{
This function launches the named class Rmd file
}
\author{
Akshay Bareja

Pol Castellano
}
