% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_slides.R
\name{launch_slides}
\alias{launch_slides}
\title{Launch Slides for each Project}
\usage{
launch_slides(project = 1, testing = FALSE)
}
\arguments{
\item{project}{Integer indicating the relevant project number.}

\item{testing}{Logical (default == FALSE). Appropriate path is selected depending on what environment the package is runing in. USERS CAN IGNORE.}
}
\value{
Opens the corresponding .Rmd file in RStudio (with or without answers).
}
\description{
This function launches the named presentation
}
\examples{
# Slides Corresponding to the Project Number 1
launch_slides(project = 1)
}
\author{
Akshay Bareja, Pol Castellano-Escuder
}
