% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_exercise.R
\name{launch_exercise}
\alias{launch_exercise}
\title{Launch .Rmd files for exercises}
\usage{
launch_exercise(exercise = 1, show_answers = FALSE, testing = FALSE)
}
\arguments{
\item{show_answers}{Logical (default == FALSE). If this parameter is set to TRUE, answers will be shown in the .Rmd file.}

\item{testing}{Logical (default == FALSE). Appropriate path is selected depending on what environment the package is runing in. USERS CAN IGNORE.}

\item{project}{Integer indicating the relevant exercise number.}
}
\value{
Opens the corresponding .Rmd file in RStudio (with or without answers).
}
\description{
This function launches the named exercise .Rmd file
}
\examples{
# Exercise Number 1
launch_exercise(exercise = 1)

# Exercise Number 1 with Answers
launch_exercise(exercise = 1, show_answers = TRUE)
}
\author{
Akshay Bareja, Pol Castellano-Escuder
}
